Fix page indices

+++ src/scrape_images.py
@@ -7,13 +7,14 @@ pages_dir = os.listdir(PATH)
 def get_image_names(path):
     image_names = None
     with open(path, 'r') as f:
-        content = f.read().strip()
-        
-        soup = BeautifulSoup(content, "html.parser")
-        elements = soup.find_all('img')
-        
-        if len(elements) > 0:
-            image_names = list(map(lambda el : el['src'], elements))
+        if path.endswith('.html'):
+            content = f.read().strip()
+            
+            soup = BeautifulSoup(content, "html.parser")
+            elements = soup.find_all('img')
+            
+            if len(elements) > 0:
+                image_names = list(map(lambda el : el['src'], elements))
         
     return image_names
 
@@ -28,12 +29,12 @@ all_images = list(set(all_images))
 
 # download all images
 def download(name):
-    print (name)
-    img_url = f'http://mcgov.co.uk/riddles/{name}'
-    # wget.download(img_url, out=f"../pages/{name}")
-    os.system(f"wget {img_url} -O ../pages/{name}")
-    # with open(f'../pages/{name}', 'wb') as f:
-        # f.write(requests.get(img_url).content)
+    if not os.path.exists(f'../pages/{name}'):
+        print (f"{name} downloading")
+        img_url = f'http://mcgov.co.uk/riddles2/{name}'
+        os.system(f"wget {img_url} -O ../pages/{name}")
+    else:
+        print (f"{name} exists")
 
 for img in all_images:
     download(img)
